ViewItem page is working  now

+++ src/components/ViewItem.vue
@@ -11,17 +11,19 @@
     <router-link to="/">Back</router-link>
     <button @click="deleteItem">Delete</button>
     <div v-if="item_id !== null">
-    <router-link :to="{name: 'edit-item', params: {item_id: item_id}}">
+      <router-link :to="{name: 'edit-item', params: {item_id: item_id}}">
         <v-btn color="red darken-2" absolute dark fab bottom right style="margin-bottom:50px">
           <v-icon>mdi-plus</v-icon>
         </v-btn>
       </router-link>
-      </div>
+    </div>
   </div>
 </template>
 
 <script>
-import db from './firebaseInit'
+import firebase from "firebase";
+import db from "./firebaseInit";
+
 export default {
   name: "view-item",
   data() {
@@ -31,36 +33,48 @@ export default {
     };
   },
   // navigation guard
-  beforeRouteEnter(to, from, next) {
-    db.collection('items').where('item_id', '==', to.params.item_id).get().then(queryShapshot => {
-      queryShapshot.forEach(doc => {
+  beforeRouteEnter(to, from, next) { // eslint-disable-line no-unused-vars
+    db.collection("users")
+      .doc(firebase.auth().currentUser.uid)
+      .collection("items")
+      .doc(to.params.item_id)
+      .get()
+      .then(querySnapshot => {
+        if (!querySnapshot.exists) {
+          next("/notFound");
+        }
+        let data = querySnapshot.data();
         next(vm => {
-          vm.item_id = doc.data().item_id
-          vm.name = doc.data().name
-        })
-      })
-    })
+          vm.name = data.name;
+        });
+      });
   },
   watch: {
-    '$route': 'fetchData'
+    $route: "fetchData"
   },
   methods: {
-    fetchData () {
-      db.collection('items').where('item_id', '==', this.$route.params.item_id).get().then(queryShapshot=>{
-        queryShapshot.forEach(doc => {
-          this.item_id = doc.data().item_id
-          this.name = doc.data().name
-        })
-      })
+    fetchData() {
+      db.collection("items")
+        .where("item_id", "==", this.$route.params.item_id)
+        .get()
+        .then(queryShapshot => {
+          queryShapshot.forEach(doc => {
+            this.item_id = doc.data().item_id;
+            this.name = doc.data().name;
+          });
+        });
     },
-    deleteItem () {
-      if (confirm('Are you sure?')) {
-         db.collection('items').where('item_id', '==', this.$route.params.item_id).get().then(queryShapshot=>{
-        queryShapshot.forEach(doc => {
-          doc.ref.delete()
-          this.$router.push('/')
-        })
-      })
+    deleteItem() {
+      if (confirm("Are you sure?")) {
+        db.collection("items")
+          .where("item_id", "==", this.$route.params.item_id)
+          .get()
+          .then(queryShapshot => {
+            queryShapshot.forEach(doc => {
+              doc.ref.delete();
+              this.$router.push("/");
+            });
+          });
       }
     }
   }
